[catcollar] Enhancement: Remove futures folder
